+++ package/nls-zh-cn.lua
@@ -11,6 +11,13 @@ return {
     ["config.runtime.nonstandardSymbol"]  = "支持非标准的符号。请务必确认你的运行环境支持这些符号。",
     ["config.runtime.plugin"]             = "（实验）插件路径。",
     ["config.runtime.fileEncoding"]       = "文件编码，`ansi` 选项只在 `Windows` 平台下有效。",
+    ['config.runtime.builtin']            = [[
+调整内置库的启用状态，你可以根据实际运行环境禁用掉不存在的库（或重新定义）。
+
+* `default`: 表示库会根据运行版本启用或禁用
+* `enable`: 总是启用
+* `disable`: 总是禁用
+]],
     ["config.diagnostics.enable"]         = "启用诊断。",
     ["config.diagnostics.disable"]        = "禁用的诊断（使用浮框括号内的代码）。\n",-- .. example.disable,
     ["config.diagnostics.globals"]        = "已定义的全局变量。\n",-- .. example.globals,